pyproject.toml
[tool.poetry]
name = "target-ai-consumer"
version = "0.1.0"
description = "AI-powered tool for diagnosing biotic stress in coffee leaves, using classification and segmentation models for analysis and reporting."
authors = ["Estevão Pimentel <estevaopbsantos@gmail.com>"]
readme = "README.md"
keywords = [
    "AI",
    "diagnosis",
    "coffee",
    "leaves",
    "classification",
    "segmentation",
]
classifiers = [
    "Development Status :: 7 - Inactive",
    "Environment :: GPU :: NVIDIA CUDA :: 12 :: 12.1",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]
repository = "https://github.com/TargetApp/target-ai-consumer"


[tool.poetry.dependencies]
python = "3.12.10"
filelock = "3.15.1"
fsspec = "2024.6.0"
Jinja2 = "3.1.4"
MarkupSafe = "2.1.5"
mpmath = "1.3.0"
mysql-connector-python = "8.4.0"
networkx = "3.3"
numpy = "1.26.4"
nvidia-cublas-cu12 = "12.1.3.1"
nvidia-cuda-cupti-cu12 = "12.1.105"
nvidia-cuda-nvrtc-cu12 = "12.1.105"
nvidia-cuda-runtime-cu12 = "12.1.105"
nvidia-cudnn-cu12 = "8.9.2.26"
nvidia-cufft-cu12 = "11.0.2.54"
nvidia-curand-cu12 = "10.3.2.106"
nvidia-cusolver-cu12 = "11.4.5.107"
nvidia-cusparse-cu12 = "12.1.0.106"
nvidia-nccl-cu12 = "2.20.5"
nvidia-nvjitlink-cu12 = "12.5.40"
nvidia-nvtx-cu12 = "12.1.105"
opencv-python = "4.10.0.82"
pillow = "10.3.0"
sympy = "1.12.1"
torch = "2.3.1"
torchvision = "0.18.1"
typing_extensions = "4.12.2"

[tool.poetry.group.dev.dependencies]
mypy = "1.15.0"
flake8 = "7.2.0"
black = "25.1.0"

[build-system]
requires = ["poetry>=1.1.12"]
build-backend = "poetry.masonry.api"

[tool.mypy]
ignore_missing_imports = true
